requirements.txt
toml
customtkinter
mysql.connector
pillow
ctkmessagebox
